Improved energy barrier calculation using E_perp
